backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2,<6.0",
    "djangorestframework>=3.14",
    "djangorestframework-simplejwt>=5.3",
    "django-cors-headers>=4.3",
    "django-filter>=23.5",
    "whitenoise>=6.6",
    "pydantic>=2.0",
    "dj-database-url>=2.0",
    "django-libsql @ git+https://github.com/Vultron-AI/django-libsql.git@main",
    "gunicorn>=21.2.0",
    "requests>=2.31",
    "PyJWT>=2.8",
    "cryptography>=42.0",
    "govector-uga-auth>=0.1.2",
    "drf-spectacular>=0.29.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0",
    "pytest-django>=4.8",
    # Linting and formatting
    "ruff>=0.14",
    "prospector>=1.10",
    "pylint-django>=2.5",
    "mypy>=1.8",
    "django-stubs>=4.2",
]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    "migrations",
    ".venv",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    "I",      # isort (import sorting)
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["."]
testpaths = ["."]
norecursedirs = [".*", "node_modules", ".venv"]
addopts = "--rootdir=."
